+++ CHANGELOG.md
@@ -8,6 +8,11 @@
  - In some rare edge cases, `igraph_pagerank()` with the ARPACK method and `igraph_hub_score()` / `igraph_authority_score()` could return incorrect results. The problem could be detected by checking that the returned eigenvalue is not negative. See issue #2090.
  - `igraph_permute_vertices()` now checks for out-of-range indices in the permutation vector.
  - `igraph_create()` now checks for non-finite vertex indices in the edges vector.
+ - `igraph_eigenvector_centrality()` would return incorrect scores when some weights were negative.
+
+### Other
+
+ - Documentation improvement.
 
 ## [0.9.9] - 2022-06-04